+++ database/demo-cn/myems_system_db.sql
@@ -260,13 +260,13 @@ INSERT INTO `myems_system_db`.`tbl_energy_flow_diagrams_links`
 (`id`, `energy_flow_diagram_id`, `source_node_id`, `target_node_id`, `meter_uuid`)
 VALUES
 (1, 1, 1, 3, '5ca47bc5-22c2-47fc-b906-33222191ea40'),
-(2, 1, 2, 4, '5d4d2f06-6200-4671-b182-4cf32cd9228f'),
-(3, 1, 2, 5, '7897665b-66ac-481d-9c31-2ab2ecbda16c'),
-(4, 1, 2, 6, 'f0c278ec-eb32-4c5e-a35f-88643b00c367'),
-(5, 1, 3, 7, '9918aa6c-79e9-4579-8f2e-a76eb9fe4e3e'),
-(6, 1, 3, 8, '831cbc8c-1429-4840-946e-f0b389b2253e'),
-(7, 1, 4, 9, 'd2fc8464-3f13-42a9-8a57-63f95f677f0f'),
-(8, 1, 4, 10, '7e4b3831-887b-40e2-b7f8-4d77c6f206a9');
+(2, 1, 2, 4, 'd6f3f56b-10ee-4d22-ad47-5acc1353a6f4'),
+(3, 1, 2, 5, '6db58cd6-33d3-58ed-a095-22333202fb51'),
+(4, 1, 2, 6, '3fff2cfb-f755-44c8-a919-6135205a8573'),
+(5, 1, 3, 7, '62f473e0-1a35-41f3-9c30-8110d75d65bb'),
+(6, 1, 3, 8, '5ca47bc5-22c2-47fc-b906-33222191ea40'),
+(7, 1, 4, 9, 'd6f3f56b-10ee-4d22-ad47-5acc1353a6f4'),
+(8, 1, 4, 10, '6db58cd6-33d3-58ed-a095-22333202fb51');
 
 COMMIT;
 
@@ -402,7 +402,7 @@ INSERT INTO `myems_system_db`.`tbl_meters`
 (`id`, `name`, `uuid`, `energy_category_id`, `is_counted`, `hourly_low_limit`, `hourly_high_limit`, `cost_center_id`, `energy_item_id`, `master_meter_id`, `description`)
 VALUES
 (1, '计量表1', '5ca47bc5-22c2-47fc-b906-33222191ea40', 1, 1, 0.000, 999.999, 1, 1, null, 'meter1'),
-(2, '计量表2', '719acf65-b932-4176-94e5-e9fe420abc68', 1, 1, 0.000, 999.999, 1, 1, 1,  'meter2'),
+(2, '计量表2', 'd6f3f56b-10ee-4d22-ad47-5acc1353a6f4', 1, 1, 0.000, 999.999, 1, 1, 1,  'meter2'),
 (3, '计量表3', '6db58cd6-33d3-58ed-a095-22333202fb51', 1, 1, 0.000, 999.999, 1, 1, 1,  'meter3');
 
 COMMIT;