+++ CHANGELOG.md
@@ -17,8 +17,10 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 - added virtual meter export, import and clone functions to myems-api, myems-admin
 - added space export, import and clone functions to myems-api
 - added command export, import and clone functions to myems-api
+- added energy flow diagram export, import and clone functions to myems-api
 ### Changed
 ### Fixed
+- fixed unkonwn meter uuid of energy flow diagram links in demo database
 ### Removed
 
 ## [v4.0.0] - 2024-01-013